fix(install): fix install script on windows to handle debug mode

+++ scripts/windows/install_google_breakpad.py
@@ -115,13 +115,19 @@ for dir in google_breakpad_library_dir_base:
 
 
 
-
+google_breakpad_binary_dir_base = [
+    'tools/windows/dump_syms/Release'
+    ]
 google_breakpad_binary_files = [
-    'tools/windows/dump_syms/Release/dump_syms.exe'
+    'dump_syms.exe'
     ]
 
-for f in google_breakpad_binary_files:
-    src, dst = convertSrcAndDst(f, os.path.basename(f), 'bin')
-    copyFile(src, dst)
+for dir in google_breakpad_binary_dir_base:
+    for buildType in build_type:
+        for file in google_breakpad_binary_files:
+            path = os.path.join(dir, buildType, 'lib', file)
+            src, dst = convertSrcAndDst(path, os.path.basename(path), 'bin')
+            if os.path.exists(src):
+                copyFile(src, dst)
